+++ Joyride/Platforms/IGesture.cs
@@ -13,5 +13,6 @@ namespace Joyride.Platforms
         Screen Swipe(Direction direction, double scale=1.0, long durationMilliSecs = 500);
         Screen Swipe(string elementName, Direction direction, double scale=1.0, long durationMilliSecs = 500);
         Screen ScrollUntil(string elementName, Direction direction, int maxRetries, int timeoutSecs, double scale = 1.0, long durationMilliSecs = 500);
+        Screen DragAndDrop(string fromElementName, string toElementName);
     }
 }