+++ Joyride/Platforms/IUiControl.cs
@@ -3,6 +3,7 @@
     public interface IUiControl
     {
         Screen EnterText(string elementName, string text);
+        Screen ClearText(string elementName);
         Screen SetCheckbox(string elementName, bool enabled = true);
     }
 }